requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-tinymce>=3.6
drf-spectacular>=0.27
requests>=2.31
Pillow>=10.0
